requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
Flask==1.1.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
python-dotenv==0.12.0
redis==3.4.1
requests==2.22.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
WTForms==2.2.1
